chore: rename swappersApi to swapeprApi
