+++ src/components/MultiHopTrade/components/TradeInput/components/TradeQuotes/TradeQuotes.tsx
@@ -9,7 +9,7 @@ import { SlideTransitionY } from 'components/SlideTransitionY'
 import { useIsSnapInstalled } from 'hooks/useIsSnapInstalled/useIsSnapInstalled'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { useWalletSupportsChain } from 'hooks/useWalletSupportsChain/useWalletSupportsChain'
-import type { ApiQuote } from 'state/apis/swappers'
+import type { ApiQuote } from 'state/apis/swapper'
 import { selectBuyAsset } from 'state/slices/selectors'
 import { selectActiveQuoteIndex } from 'state/slices/tradeQuoteSlice/selectors'
 import { useAppSelector } from 'state/store'
