+++ src/components/MultiHopTrade/hooks/useSupportedAssets.tsx
@@ -4,7 +4,7 @@ import { useIsSnapInstalled } from 'hooks/useIsSnapInstalled/useIsSnapInstalled'
 import { useWallet } from 'hooks/useWallet/useWallet'
 import { walletSupportsChain } from 'hooks/useWalletSupportsChain/useWalletSupportsChain'
 import { isSome } from 'lib/utils'
-import { useGetSupportedAssetsQuery } from 'state/apis/swappers/swappersApi'
+import { useGetSupportedAssetsQuery } from 'state/apis/swapper/swapperApi'
 import { selectAssetsSortedByMarketCapUserCurrencyBalanceAndName } from 'state/slices/common-selectors'
 import { selectAssets } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
