+++ src/state/slices/tradeQuoteSlice/selectors.ts
@@ -7,9 +7,9 @@ import { getDefaultSlippageDecimalPercentageForSwapper } from 'constants/constan
 import type { Selector } from 'reselect'
 import { bn, bnOrZero } from 'lib/bignumber/bignumber'
 import { fromBaseUnit } from 'lib/math'
-import type { ApiQuote, ErrorWithMeta, TradeQuoteError } from 'state/apis/swappers'
-import { TradeQuoteWarning } from 'state/apis/swappers'
-import { selectSwappersApiTradeQuotes } from 'state/apis/swappers/selectors'
+import type { ApiQuote, ErrorWithMeta, TradeQuoteError } from 'state/apis/swapper'
+import { TradeQuoteWarning } from 'state/apis/swapper'
+import { selectSwappersApiTradeQuotes } from 'state/apis/swapper/selectors'
 import { isCrossAccountTradeSupported } from 'state/helpers'
 import type { ReduxState } from 'state/reducer'
 import { createDeepEqualOutputSelector } from 'state/selector-utils'
